+++ src/cores/dynamo/dynamo-scan-service.ts
@@ -8,8 +8,8 @@
  *
  * @copyright (C) 2020 LemonCloud Co Ltd. - All Rights Reserved.
  */
-import { _log, _inf, $U } from '../engine/';
-import { GeneralItem } from './core-types';
+import { _log, _inf, $U } from '../../engine/';
+import { GeneralItem } from '../core-types';
 import { DynamoOption, DynamoService } from './dynamo-service';
 const NS = $U.NS('DSCN', 'green'); // NAMESPACE TO BE PRINTED.
 
@@ -92,8 +92,8 @@ export interface DynamoSimpleScannable<T extends GeneralItem> {
 /** ****************************************************************************************************************
  *  Service Main
  ** ****************************************************************************************************************/
-import Scan from '../lib/dynamo/scan';
-import Serializer from '../lib/dynamo/serializer';
+import Scan from '../../lib/dynamo/scan';
+import Serializer from '../../lib/dynamo/serializer';
 
 /**
  * class: `DynamoScanService`
