+++ src/cores/dynamo/index.ts
@@ -1,0 +1,12 @@
+/**
+ * `cores/dynamo/index.tx`
+ * - default index to export
+ *
+ * @author      Steve Jung <steve@lemoncloud.io>
+ * @date        2022-05-13 initial file to group
+ *
+ * @copyright (C) lemoncloud.io 2022 - All Rights Reserved.
+ */
+export * from './dynamo-service';
+export * from './dynamo-query-service';
+export * from './dynamo-scan-service';
